* HierarchyRebuildContext refactoring
