+++ com.dotsui.core/DotsUI.Core/HierarchyRebuildContext.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 00326afc96caba743ba6a62bcb009cf4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
